Create parent directories for target container file in case they do not exist
When packaging=pom is used it might be that the 'target' directory is not created. The plugin then tries to filter the Containerfile and store it in the target dir. However, that directory does not exist and the copyFile fails with 'Failed to filter Containerfile! <path> (No such file or directory).'

This fixes #51